requirements.txt
fastapi
uvicorn
supabase
python-jose[cryptography]
passlib[bcrypt]
streamlit
requests
cryptography
pqcrypto
langchain
langchain-groq
scikit-learn
pandas
python-dotenv
boto3
pydantic
pydantic-settings
python-multipart
numpy
